Various and sundry cleanup tasks
Delete unused things. Move some things out of core files and into
modules that indicate that they are test only. Move script_relative_path
to utils rather than test because it useful outside of tests.
