+++ python_modules/dagster-ge/dagster_ge_tests/test_pandas_ge.py
@@ -2,19 +2,21 @@ import pytest
 
 import pandas as pd
 
-import dagster
-import dagster.pandas_kernel as dagster_pd
-import dagster_ge
 from dagster import (
-    config,
-    solid,
-    execute_pipeline,
+    DependencyDefinition,
     InputDefinition,
     OutputDefinition,
-    DependencyDefinition,
+    PipelineDefinition,
+    config,
+    execute_pipeline,
+    solid,
 )
-from dagster.core.errors import (DagsterExecutionFailureReason, DagsterExpectationFailedError)
-from dagster.utils.test import script_relative_path
+from dagster.core.errors import DagsterExpectationFailedError
+from dagster.core.utility_solids import define_stub_solid
+import dagster.pandas_kernel as dagster_pd
+from dagster.utils import script_relative_path
+
+import dagster_ge
 
 
 def _sum_solid_impl(num_df):
@@ -71,12 +73,9 @@ def sum_solid_expectations_config(num_df):
     return _sum_solid_impl(num_df)
 
 
-from dagster.core.utility_solids import define_stub_solid
-
-
 def test_single_node_passing_expectation():
     in_df = pd.DataFrame.from_dict({'num1': [1, 3], 'num2': [2, 4]})
-    pipeline = dagster.PipelineDefinition(
+    pipeline = PipelineDefinition(
         solids=[define_stub_solid('value', in_df), sum_solid],
         dependencies={'sum_solid': {
             'num_df': DependencyDefinition('value')
@@ -96,7 +95,7 @@ def test_single_node_passing_expectation():
 
 def test_single_node_passing_json_config_expectations():
     in_df = pd.DataFrame.from_dict({'num1': [1, 3], 'num2': [2, 4]})
-    pipeline = dagster.PipelineDefinition(
+    pipeline = PipelineDefinition(
         solids=[define_stub_solid('value', in_df), sum_solid_expectations_config],
         dependencies={
             sum_solid_expectations_config.name: {
@@ -118,7 +117,7 @@ def test_single_node_passing_json_config_expectations():
 
 def test_single_node_failing_expectation():
     in_df = pd.DataFrame.from_dict({'num1': [1, 3], 'num2': [2, 4]})
-    pipeline = dagster.PipelineDefinition(
+    pipeline = PipelineDefinition(
         solids=[define_stub_solid('value', in_df), sum_solid_fails_input_expectation],
         dependencies={
             sum_solid_fails_input_expectation.name: {
