+++ python_modules/dagster/dagster/pandas_kernel/pandas_kernel_tests/test_pandas_hello_world_no_library_slide.py
@@ -14,7 +14,7 @@ from dagster import (
     types,
 )
 
-from dagster.utils.test import script_relative_path
+from dagster.utils import script_relative_path
 
 
 def define_read_csv_solid(name):
