+++ python_modules/dagster/dagster/utils/__init__.py
@@ -1,30 +1,16 @@
-from __future__ import (absolute_import, division, print_function, unicode_literals)
-from builtins import *  # pylint: disable=W0622,W0401
 import inspect
+import os
 
 from dagster import check
+def script_relative_path(file_path):
+    '''
+    Useful for testing with local files. Use a path relative to where the
+    test resides and this function will return the absolute path
+    of that file. Otherwise it will be relative to script that
+    ran the test
+    '''
+    # from http://bit.ly/2snyC6s
 
-
-def has_context_argument(fn):
-    check.callable_param(fn, 'fn')
-
-    argspec = inspect.getfullargspec(fn)
-    return 'context' in argspec[0]
-
-
-def make_context_arg_optional(fn):
-    check.callable_param(fn, 'fn')
-    # make this idempotent during the transition
-    if fn.__name__ == '__wrapper_with_context':
-        return fn
-    check.invariant(fn.__name__ != '__wrapper_with_context')
-
-    if not has_context_argument(fn):
-
-        def __wrapper_with_context(*args, context, **kwargs):
-            check.not_none_param(context, 'context')
-            return fn(*args, **kwargs)
-
-        return __wrapper_with_context
-    else:
-        return fn
+    check.str_param(file_path, 'file_path')
+    scriptdir = inspect.stack()[1][1]
+    return os.path.join(os.path.dirname(os.path.abspath(scriptdir)), file_path)
