+++ python_modules/dagster/dagster/utils/test.py
@@ -1,4 +1,3 @@
-import inspect
 import itertools
 import os
 import tempfile
@@ -7,18 +6,6 @@ from contextlib import contextmanager
 from dagster import check
 
 
-def script_relative_path(file_path):
-    '''
-    Useful for testing with local files. Use a path relative to where the
-    test resides and this function will return the absolute path
-    of that file. Otherwise it will be relative to script that
-    ran the test
-    '''
-    # from http://bit.ly/2snyC6s
-
-    check.str_param(file_path, 'file_path')
-    scriptdir = inspect.stack()[1][1]
-    return os.path.join(os.path.dirname(os.path.abspath(scriptdir)), file_path)
 
 
 def _unlink_swallow_errors(path):
