+++ python_modules/dagster/dagster_tests/test_config.py
@@ -1,4 +1,3 @@
-import pytest
 import yaml
 
 from dagster import config
